+++ Changelog.md
@@ -1,3 +1,6 @@
+#### 0.1.0
+ - Error first convention for `loadMainFile` and other jxcore calls (Breaking change)
+
 #### 0.0.8 - DEV Version
  - Update to latest JXcore 0.3.0.7
  - Fixes Android WiFi reconnection issue